Pytorch2onnx which can be used for DIM and GCA model (#105)
* add pytorch2onnx for DIM

* support convert to ONNX for GCA

* use forward_dummy for onnx export

* remove useless comment

* retrieve useless modify

* Updating according to the latest mmcv

* modify the comment

* add docstring

* Fix docstring dot

+++ mmedit/models/common/gca_module.py
@@ -341,8 +341,10 @@ class GCAModule(nn.Module):
 
     def get_self_correlation_mask(self, img_feat):
         _, _, h, w = img_feat.shape
+        # As ONNX does not support dynamic num_classes, we have to convert it
+        # into an integer
         self_mask = F.one_hot(
-            torch.arange(h * w).view(h, w), num_classes=h * w)
+            torch.arange(h * w).view(h, w), num_classes=int(h * w))
         self_mask = self_mask.permute(2, 0, 1).view(1, h * w, h, w)
         # use large negative value to mask out self-correlation before softmax
         self_mask = self_mask * self.penalty
